+++ include/cantera/zeroD/MoleReactor.h
@@ -47,7 +47,7 @@ protected:
     virtual void getSurfaceInitialConditions(double* y);
 
     //! const value for the species start index
-    const int m_sidx = 2;
+    const size_t m_sidx = 2;
 };
 
 }
